Adds first attempt at Chapter 1, Interview Question 1.
With the lack of information (for example, answers from an interviewer to
questions from me regarding the problem at hand), I had to make some
assumptions... and.. well, it's...

I should figure out a better way. But at least I have a collection of test
cases to use for testing now.